--- WeCantSpell.Hunspell.Tests/WeCantSpell.Hunspell.Tests.net48.v3.ncrunchproject
@@ -1,5 +1,0 @@
-﻿<ProjectConfiguration>
-  <Settings>
-    <EnableRDI>False</EnableRDI>
-  </Settings>
-</ProjectConfiguration>
