--- WeCantSpell.Hunspell/WeCantSpell.Hunspell.netstandard2.0.v3.ncrunchproject
@@ -1,5 +1,0 @@
-﻿<ProjectConfiguration>
-  <Settings>
-    <EnableRDI>False</EnableRDI>
-  </Settings>
-</ProjectConfiguration>
